requirements.prod.txt
torch==2.6.0+cu124
torchvision==0.21.0+cu124

accelerate==1.9.0
diffusers==0.34.0
transformers==4.54.1
huggingface-hub==0.34.1

numpy==2.2.0
scipy==1.16.1
einops==0.8.1
omegaconf==2.3.0
tqdm==4.67.1

pillow==11.3.0
opencv-python-headless==4.12.0.88
imageio==2.37.0
imageio-ffmpeg==0.6.0

rembg==2.0.67
PyMatting==1.1.14

gradio==5.39.0

trimesh==4.7.1
xatlas==0.0.11

-e git+https://github.com/NVlabs/nvdiffrast.git@729261dc64c4241ea36efda84fbf532cc8b425b8#egg=nvdiffrast
